fix fallout

+++ src/libcore/num/f64.rs
@@ -165,7 +165,7 @@ impl Float for f64 {
     /// Returns `true` if the number is infinite.
     #[inline]
     fn is_infinite(self) -> bool {
-        self == Float::infinity() || self == Float::neg_infinity()
+        self == f64::infinity() || self == f64::neg_infinity()
     }
 
     /// Returns `true` if the number is neither infinite or NaN.
@@ -322,14 +322,14 @@ impl Float for f64 {
     /// `Float::infinity()`.
     #[inline]
     fn is_positive(self) -> bool {
-        self > 0.0 || (1.0 / self) == Float::infinity()
+        self > 0.0 || (1.0 / self) == f64::infinity()
     }
 
     /// Returns `true` if `self` is negative, including `-0.0` and
     /// `Float::neg_infinity()`.
     #[inline]
     fn is_negative(self) -> bool {
-        self < 0.0 || (1.0 / self) == Float::neg_infinity()
+        self < 0.0 || (1.0 / self) == f64::neg_infinity()
     }
 
     /// Fused multiply-add. Computes `(self * a) + b` with only one rounding
